chore: load permissions from request
