+++ advanced/wallets/react-wallet-v2/src/utils/EIP155RequestHandlerUtil.ts
@@ -75,7 +75,7 @@ export async function approveEIP155Request(requestEvent: RequestEventArgs) {
         
         // intercept for smart account getPermissions mock
         if(domain.name === 'eth_getPermissions_v1' && wallet instanceof KernelSmartAccountLib){
-          const sessionKey = await wallet.issueSessionKey(data.targetAddress)
+          const sessionKey = await wallet.issueSessionKey(data.targetAddress,data.permissions)
           return formatJsonRpcResult(id, sessionKey)
         }
